Updated Server and Password Hashing
I removed the old test driver and replaced it with the new one, almost the same but just a few new tests and a new file name. The Password hashing class is a class that is required to hash out a password using SHA-256 in order to create some password security over the server. The server itself was updated to include the hashing as well as it now contains the ability for a user to add or remove their favourite movies to their profile and a minor adjustment for event movies to lose a vote.